+++ flutter/CHANGELOG.md
@@ -1,11 +1,14 @@
 ## [0.0.4] - 8/9 2020
 updated dependencies, updated import & usage patterns
 use as like so.
-
 ```dart
 import 'package:flutter_remote_ui/flutter_remote_ui.dart' as rui;
 // rui. ....
 ```
+
+-----
++ added route row as prebuilt widget
+
 
 ## [0.0.3] - 6/26 2020
 remote icons now available remote ui - remote icon